Move form to separate file
