+++ src/components/features/UniversalForm/UniversalForm.module.scss
@@ -1,0 +1,14 @@
+@import '../../../styles/settings.scss';
+
+.component {
+  font-family: Roboto;
+  color: grey;
+  text-align: center;
+  max-width: 500px;
+  margin: auto;
+}
+
+.input {
+  width: 200px;
+  margin: 20px;
+}
